Provide hit/combo channels per player; add player ids.
Now the two players have their own channels for registering hits and
combos; this prevents the inevitable confusion if someone racks up a
combo and then we switch players before it is finalized.

Also added a simple `:player-id` key to each player, which may be
`:player-one` or `:player-two`. This lets us do a quick and dirty player
comparison even when cursor state is inconsistent. Or appears to be
inconsistent -- I'm sure my difficulty is founded in my own lack of
understanding right now.