Position and style combo damage view. Alpha quality.
The combo damage view is now in more or less the right place, on top of
the seam between character health bars. When no combo is in progress, it
reads "VS" and has a green background; when a combo is in progress, it
displays the running total, and has a red background.