Add :preamble to advanced profile to enable compilation.
According to the Om documentation, it's necessary to define both
`:preamble ["react/react.min.js"]` and `:externs
["react/externs/react.js"]`. In practice, I found that adding the
`:externs` declaration caused a "duplicate extern" error from the
Closure compiler. Maybe the React file is self-externing? It comes from
a jar pulled in as an Om dependency, so maybe it has some extra niceties
specifically for ClojureScript.