Temporarily hack index.html to use advanced-compile JS file.
After advanced compilation, there is a single JS file which contains
everything but fastclick.js. We can just include it and let it work.

In the dev build, however, we have separate script tags for goog/base
and our own code, and then we call `goog.require` on our core namespace
to run the app. Better for debugging.

For now, we'll just comment out the version we don't want. Soon we want
build-specific HTML files, so we don't have to keep changing this one.