Track current player by id, not by value copy.
Observed a bug where damaging a player caused that player to lose
"current" status. I realized that we were comparing the player to
(:current-player app), but then we expected equality comparisons to work
even after a player had been damaged. This would be true in languages
with reference-based mutable state, but not in ClojureScript!

Changed (:current-player app) to just :current-player-id, and began
comparing on that.