Reset to initial data state when returning to character select.
The initial data state is now held in `alpha-counter.data/initial-data`.
When we wish to return to character select, we transact on the app state
to apply each key-value pair from initial-data to app-state.

Renamed `return-to-character-select!` to `reset-app!`. The current
implementation is a bit crunchy, but it gets the job done.